Refactor packages scounts>team2 and lab1myfirstapp>chatapp Instances of lab1myfirstapp throughout all changed to some variant of chatapp
