+++ .idea/.name
@@ -1,1 +1,1 @@
-Lab 1 My First App
+Chat App
