+++ .idea/misc.xml
@@ -3,6 +3,8 @@
   <component name="DesignSurface">
     <option name="filePathToZoomLevelMap">
       <map>
+        <entry key="..\:/Users/Nicol/AndroidStudioProjects/Team-2-TCSS-450/app/src/main/res/layout/activity_main.xml" value="0.24818840579710144" />
+        <entry key="..\:/Users/Nicol/AndroidStudioProjects/Team-2-TCSS-450/app/src/main/res/layout/fragment_first.xml" value="0.24818840579710144" />
         <entry key="..\:/Users/Sierra/AndroidStudioProjects/Lab1MyFirstApp/app/src/main/res/layout/activity_main.xml" value="0.31666666666666665" />
         <entry key="..\:/Users/Sierra/AndroidStudioProjects/Lab1MyFirstApp/app/src/main/res/layout/fragment_first.xml" value="0.3307291666666667" />
         <entry key="..\:/Users/Sierra/AndroidStudioProjects/Lab1MyFirstApp/app/src/main/res/layout/fragment_second.xml" value="0.32708333333333334" />
