+++ app/build.gradle
@@ -6,7 +6,7 @@ android {
     compileSdk 31
 
     defaultConfig {
-        applicationId "edu.uw.tcss450.scounts.lab1myfirstapp"
+        applicationId "edu.uw.tcss450.team2.chatapp"
         minSdk 24
         targetSdk 31
         versionCode 1
