+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="edu.uw.tcss450.scounts.lab1myfirstapp">
+    package="edu.uw.tcss450.team2.chatapp">
 
     <application
         android:allowBackup="true"
@@ -8,9 +8,9 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.Lab1MyFirstApp">
+        android:theme="@style/Theme.ChatApp">
         <activity
-            android:name=".MainActivity"
+            android:name="edu.uw.tcss450.team2.chatapp.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
