+++ app/src/main/java/edu/uw/tcss450/team2/chatapp/MainActivity.java
@@ -1,4 +1,4 @@
-package edu.uw.tcss450.scounts.lab1myfirstapp;
+package edu.uw.tcss450.team2.chatapp;
 
 import androidx.appcompat.app.AppCompatActivity;
