+++ app/src/main/java/edu/uw/tcss450/team2/chatapp/ui/FirstFragment.java
@@ -1,4 +1,4 @@
-package edu.uw.tcss450.scounts.lab1myfirstapp.ui;
+package edu.uw.tcss450.team2.chatapp.ui;
 
 import android.os.Bundle;
 
@@ -12,8 +12,7 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
-import edu.uw.tcss450.scounts.lab1myfirstapp.R;
-import edu.uw.tcss450.scounts.lab1myfirstapp.databinding.FragmentFirstBinding;
+import edu.uw.tcss450.team2.chatapp.databinding.FragmentFirstBinding;
 
 /**
  * A simple {@link Fragment} subclass.
