+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/firstFragment"
-        android:name="edu.uw.tcss450.scounts.lab1myfirstapp.ui.FirstFragment"
+        android:name="edu.uw.tcss450.team2.chatapp.ui.FirstFragment"
         android:label="fragment_first"
         tools:layout="@layout/fragment_first" >
         <action
@@ -16,7 +16,7 @@
     </fragment>
     <fragment
         android:id="@+id/secondFragment"
-        android:name="edu.uw.tcss450.scounts.lab1myfirstapp.ui.SecondFragment"
+        android:name="edu.uw.tcss450.team2.chatapp.ui.SecondFragment"
         android:label="fragment_second"
         tools:layout="@layout/fragment_second" >
         <argument
