+++ settings.gradle
@@ -6,5 +6,5 @@ dependencyResolutionManagement {
         jcenter() // Warning: this repository is going to shut down soon
     }
 }
-rootProject.name = "Lab 1 My First App"
+rootProject.name = "Chat App"
 include ':app'
